+++ bootstraptest/test_ractor.rb
@@ -148,25 +148,25 @@ assert_equal '[:ok, :ok, :ok]', %q{
 # Ractor.make_shareable issue for locals in proc [Bug #18023]
 assert_equal '[:a, :b, :c, :d, :e]', %q{
   v1, v2, v3, v4, v5 = :a, :b, :c, :d, :e
-  closure = Ractor.current.instance_eval{ Proc.new { [v1, v2, v3, v4, v5] } }
+  closure = Ractor.shareable_lambda { [v1, v2, v3, v4, v5] }
 
-  Ractor.make_shareable(closure).call
+  closure.call
 }
 
 # Ractor.make_shareable issue for locals in proc [Bug #18023]
 assert_equal '[:a, :b, :c, :d, :e, :f, :g]', %q{
   a = :a
-  closure = Ractor.current.instance_eval do
-    -> {
+  closure = Ractor.shareable_lambda do
+    Ractor.shareable_lambda {
       b, c, d = :b, :c, :d
-      -> {
+      Ractor.shareable_lambda {
         e, f, g = :e, :f, :g
-        -> { [a, b, c, d, e, f, g] }
+        Ractor.shareable_lambda { [a, b, c, d, e, f, g] }
       }.call
     }.call
-  end
+  end.call
 
-  Ractor.make_shareable(closure).call
+  closure.call
 }
 
 ###
@@ -967,7 +967,7 @@ assert_equal 'can not access non-shareable objects in constant C::CONST by non-m
   end
   RUBY
 
-# Constant cache should care about non-sharable constants
+# Constant cache should care about non-shareable constants
 assert_equal "can not access non-shareable objects in constant Object::STR by non-main Ractor.", <<~'RUBY', frozen_string_literal: false
   STR = "hello"
   def str; STR; end
@@ -1137,41 +1137,17 @@ assert_equal 'true', %q{
   [a.frozen?, a[0].frozen?] == [true, false]
 }
 
-# Ractor.make_shareable(a_proc) makes a proc shareable.
+# Ractor.make_shareable(a_proc) is not supported now.
 assert_equal 'true', %q{
-  a = [1, [2, 3], {a: "4"}]
+  pr = Proc.new{}
 
-  pr = Ractor.current.instance_eval do
-    Proc.new do
-      a
-    end
+  begin
+    Ractor.make_shareable(pr)
+  rescue Ractor::Error
+    true
+  else
+    false
   end
-
-  Ractor.make_shareable(a) # referred value should be shareable
-  Ractor.make_shareable(pr)
-  Ractor.shareable?(pr)
-}
-
-# Ractor.make_shareable(a_proc) makes inner structure shareable and freezes it
-assert_equal 'true,true,true,true', %q{
-  class Proc
-    attr_reader :obj
-    def initialize
-      @obj = Object.new
-    end
-  end
-
-  pr = Ractor.current.instance_eval do
-    Proc.new {}
-  end
-
-  results = []
-  Ractor.make_shareable(pr)
-  results << Ractor.shareable?(pr)
-  results << pr.frozen?
-  results << Ractor.shareable?(pr.obj)
-  results << pr.obj.frozen?
-  results.map(&:to_s).join(',')
 }
 
 # Ractor.shareable?(recursive_objects)
@@ -1203,7 +1179,7 @@ assert_equal '[C, M]', %q{
 }
 
 # Ractor.make_shareable with curried proc checks isolation of original proc
-assert_equal 'isolation error', %q{
+assert_equal 'true', %q{
   a = Object.new
   orig = proc { a }
   curried = orig.curry
@@ -1212,6 +1188,8 @@ assert_equal 'isolation error', %q{
     Ractor.make_shareable(curried)
   rescue Ractor::IsolationError
     'isolation error'
+  rescue Ractor::Error => e
+    e.message.match?(/can not make shareable object/)
   else
     'no error'
   end
@@ -1221,7 +1199,7 @@ assert_equal 'isolation error', %q{
 assert_equal '1', %q{
   class C
     a = 1
-    define_method "foo", Ractor.make_shareable(Proc.new{ a })
+    define_method "foo", Ractor.shareable_proc{a}
     a = 2
   end
 
@@ -1230,17 +1208,13 @@ assert_equal '1', %q{
 
 # Ractor.make_shareable(a_proc) makes a proc shareable.
 assert_equal 'can not make a Proc shareable because it accesses outer variables (a).', %q{
-  a = b = nil
-  pr = Ractor.current.instance_eval do
-    Proc.new do
+  begin
+    a = b = nil
+    pr = Ractor.shareable_proc do
       c = b # assign to a is okay because c is block local variable
       # reading b is okay
       a = b # assign to a is not allowed #=> Ractor::Error
     end
-  end
-
-  begin
-    Ractor.make_shareable(pr)
   rescue => e
     e.message
   end
@@ -1462,42 +1436,6 @@ assert_equal "ok", %q{
   1_000.times { port.receive }
   "ok"
 } if !yjit_enabled? && ENV['GITHUB_WORKFLOW'] != 'ModGC' # flaky
-
-assert_equal "ok", %q{
-  def foo(*); ->{ super }; end
-  begin
-    Ractor.make_shareable(foo)
-  rescue Ractor::IsolationError
-    "ok"
-  end
-}
-
-assert_equal "ok", %q{
-  def foo(**); ->{ super }; end
-  begin
-    Ractor.make_shareable(foo)
-  rescue Ractor::IsolationError
-    "ok"
-  end
-}
-
-assert_equal "ok", %q{
-  def foo(...); ->{ super }; end
-  begin
-    Ractor.make_shareable(foo)
-  rescue Ractor::IsolationError
-    "ok"
-  end
-}
-
-assert_equal "ok", %q{
-  def foo((x), (y)); ->{ super }; end
-  begin
-    Ractor.make_shareable(foo([], []))
-  rescue Ractor::IsolationError
-    "ok"
-  end
-}
 
 # check method cache invalidation
 assert_equal "ok", %q{